refactor #22: 피드백 반영
1. AlertPopup 확장해서 이미지 추가
2.  0/10 부분의 텍스트 색상 수정
