+++ dogether/Presentation/Features/Main/MainViewModel.swift
@@ -106,7 +106,7 @@ extension MainViewModel {
     private func updateListInfo(updateList: @escaping () -> Void) {
         Task {
             let (date, status) = try await mainUseCase.getTodosInfo(dateOffset: dateOffset, currentFilter: currentFilter)
-//            todoList = try await challengeGroupsUseCase.getMyTodos(date: date, status: status) // ✅
+            todoList = try await challengeGroupsUseCase.getMyTodos(date: date, status: status)
             mainViewStatus = currentFilter == .all && todoList.isEmpty ? .emptyList : .todoList
             isBlockPanGesture = await todoListHeight < Int(UIScreen.main.bounds.height - (SheetStatus.normal.offset + 140 + 48))
             await MainActor.run { updateList() }
@@ -122,7 +122,7 @@ extension MainViewModel {
     
     func updateListInfo() async throws {
         let (date, status) = try await mainUseCase.getTodosInfo(dateOffset: dateOffset, currentFilter: currentFilter)
-//        todoList = try await challengeGroupsUseCase.getMyTodos(date: date, status: status) // ✅
+        todoList = try await challengeGroupsUseCase.getMyTodos(date: date, status: status)
         mainViewStatus = currentFilter == .all && todoList.isEmpty ? .emptyList : .todoList
         isBlockPanGesture = await todoListHeight < Int(UIScreen.main.bounds.height - (SheetStatus.normal.offset + 140 + 48))
     }
